mu4e: set symbol prop definition-name in defun macros
This allows `find-function' to find the definition.  While the
definition doesn't contain much useful information jumping there
instead of the beginning of the file is still better because the macro
used to define them is defined right above.

Also remove the comments about wanting to define the commands in a
loop.  One shouldn't do that; it would again make it impossible to
find the definition.