mu4e: autoload major entry points
Autoload functions `mu4e' and `mu4e~compose-mail' and the definition of
the mail-user-agent `mu4e-user-agent'.  The last to are required so
that `compose-mail' can be used without explicitly loading mu4e.

Also require `mu4e' in mu4e-compose.el (after providing mu4e-compose
to avoid a circular dependencies).  This is required because the
autoload definitions only cause mu4e-compose but not mu4e to be loaded
when using `compose-mail', but if mu4e isn't loaded various handlers
are not loaded (and possibly there are other problems if mu4e is not
fully loaded).