Automatically scroll the mail fetch status window
My workflow is to run the fetch process on demand without an update
interval. In my case, I use offlineimap. This pops up the status window,
which shows only the top-most lines of offlineimap output. I made these
changes to auto-scroll the window so that I can get a feel for what
offlineimap is up to.

Since I made message insertion more clever, and password prompting was
unrelated, I unnested it. I then added auto-scrolling to the status
window. Finally, I found that when the status window was selected, point
would move out from under me to the end of the buffer. Changing
save-current-buffer to save-excursion left point alone.

To test this, I ran M-u from *mu4e-headers* and watched the window
scroll automatically. Then I ran it with the window selected and
the automatic scrolling stopped until I deselected the window.