mu4e: fix unread face property decision
When un unread message contained attachment or was signed, it was not
coloured with unread face in the headers buffer, because the face decision
process was based upon the first flag of the message, while the unread
flag usually came last.

This commit fixes the problem by simply looking at the last flag instead
of the first flag when taking the face attribution decision.  In this way
all unread messages are preferentially given the unread face regardless
of other flags they may have.  (Simple but works; could be further improved
by not assuming any flag order.)