Consider X-Original-To a contact field
Some popular mail systems add an X-Original-To header to delivered messages so that recipients of BCC'd mails can identify the address to which the mail was sent. Postfix is one such system:

http://www.postfix.org/virtual.8.html

This patch allows mu4e-message-contact-field-matches to match contacts within the X-Original-To field, which allows proper refiling of such messages.