Add a customisation option for completion system
Introduced mu4e-use-native-completing-read customisation setting to
control which type of minibuffer completion system mu4e uses.

If unset (the default), then ido-completing-read is used, which provides
an intuitive completion system even if users don't customise mu4e
further.

If this is set to 't' then mu4e uses the standard Emacs completing-read
function for this. This can be chosen if the user prefers the native
method (perhaps unlikely), or else if they have another minibuffer
completion system that overrides completing-read and would prefer to use
that. An example of that is helm, which can override completing-read
if the helm-mode function is called.