Allow rendering to be chosen based on body-part
I added the alist, mu4e-view-render-func-alist, to capture the part to
rendering function mapping.  I moved the HTML stripping code from
mu4e-message into mu4e-view so that the part and text choosing code
could be shared among the rendering functions.  Hopefully this will not
disrupt anyone.

The function that chooses the part to render will return a cons
of the part-name and part-text and is called more that once, at present.
Once for choosing which part to render, and then again to do the actual
rendering.  The reason for this is that I like that the rendering
function takes the msg directly as it leaves it totally open to render
it how it chooses, but it is a bit less than optimal.