mu4e-actions.el: make git apply patch history aware
As you are often applying a series of patches to a tree it doesn't
make sense to move the user to default-directory every time. This
changes read-directory-name to use ido-read-directory-name and
populate the directory with the last one applied.

Any path chosen is also added to the beginning of
ido-work-directory-list for next time.