Added hook called when about to update mail
Added new hook 'mu4e-pre-update-mail-hook' that is called when mu4e is
about to update mail (via mu4e-update-mail-and-index or with the 'U'
command). This allows the hook function to make last-minute
customisations before mail is collected.

One use for this is to change the contents of mu4e-get-mail-command so
that mu4e can perform a quick mail collection most of the time by
choosing a folder subset, but periodically collect mail from all
folders.