mu-find(1): Document that you can search throug body:*
This is a long-standing feature, it works, but there was no mention of
it in the docs.

There's also bodyhtml:* but that doesn't seem to do anything for me,
maybe I just don't have enough HTML mail.

These also exist:

    path,l          ???
    refs,r          ???

But I don't know what they're for. They should probably be documented
too.