mu4e-mark: Add fancy characters for trash / delete
As threatened in my last commit. I found this via
https://en.wikipedia.org/wiki/Miscellaneous_Symbols_and_Pictographs &
http://www.unicode.org/charts/PDF/U1F300.pdf

I could never remember which thing d and D were, this makes it easier to
remember that. I split off this patch because maybe this doesn't belong
in mu4e since some might not like the characters, but I just wanted to
be able to have this customizable.