Partially solve issue #75 & add a :report-processed-every argument
This argument can be report-processed-every for report no progress, or N
where N >1 for report every N. In mu4e we just use this to pass 0 if
`mu4e-hide-index-messages' is t. This makes the protocol a lot less
chatty during indexing.

We could add something to make the "send us an update every N" to mu4e,
that's supported by mu-server now, but I'm not doing that, this solves
my use-case of having a less chatty *mu4e-log*.

I was trying to do something with (apply) and (mapconcat) in
mu4e~proc-index that would make the if/else branches less copy/pasted,
but couldn't come up with something that wasn't even more verbose and
unwieldy.

This resolves issue #715 and issue #716.