Add mu4e-compose-post-hook
This new hook allows the user to specify an action to take place after
compose buffer has been created & major mode set, but does not attach
the hook to the major mode setting. This simplifies the use of
org-mu4e-org-compose-mode, and hopefully will bridge the gap to a more
stable replacementt for that pseudo minor mode.