mu4e: smarter initialization of mu4e-msg2pdf
Before this commit the initialization of the mu4e-msg2pdf variable
assumed that the build directory was available. However, that's not the
case when installing mu from a distro package (in my case from OBS).

With this change, mu4e will look for the `msg2pdf' binary on the
`exec-path' first. If the binary exists, then this will be the initial
value. If the binary doesn't exist, then we fallback to the
`mu4e-builddir' approach.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>