mu4e: rework html2text conversion
Rework the conversion from html to text; the functions doing that now
expect to receive one parameter, the message, and return the converted
message.

In the old way, the function got invoked in a buffer with html text, and
were expected to modify it to text. This old way is still supported for
backward compatibility.