mu4e: strip extraneous newline on draft save
The removal of the separator line did not remove the `\n` introduced by the
separator insertion, so every message started with an additional newline.

Adds unit tests for the function, which can be run as follows:

```
(require 'mu4e-tests)
(ert t)
```

`ert` an Emacs-builtin.

The function names changed earlier, update the corresponding docs as well.