Fix mu4e-view-mode-hook not called on unread messages
Fix #1265 and #1267.

Here's the problem:

When `mu4e~view-internal` is called it makes a call to
`mu4e~view-mark-as-read-maybe`.
When the message is unread `mu4e~view-mark-as-read-maybe` calls
`mu4e~proc-move` which flag the message as read and calls again
`mu4e~view-internal` which in turn calls again
`mu4e~view-mark-as-read-maybe` which this time returns nil and
`mu4e~view-internal` continues printing the view buffer and running
hooks.
Then `mu4e~view-mark-as-read-maybe` first call continue to run and
calls now `mu4e~proc-view` which as `mu4e~proc-move` calls again
`mu4e~view-internal` and prints a second time the buffer and run a
second time the hooks.

The first thing this patch does is always running hooks after printing
message, then to avoid doing this again it advice temporarily
`mu4e-view-func` aka `mu4e~view-internal` to prevent printing and
running hooks again.
Finally this advice function remove itself the advice to allow
printing the next message.

* mu4e/mu4e-view.el (mu4e~view-internal): Remove mode-enabled checking
which is useless.
(mu4e~view-advice-view-func): Advice mu4e-view-func to ensure it
doesn't run a second time.
(mu4e~view-mark-as-read-maybe): Use it.