Add custom variable mu4e-hide-main-view-buffer
If non-nil, mu4e's main view buffer will be replaced with a minibuffer
prompt.

The function `mu4e-main-view` is modified to rely on this variable
rather than `mu4e-split-view` (currently, the minibuffer is used if
and only if `mu4e-split-view` is set to `'single-window`. Thus one can
now use single-window mode together with the main view buffer (useful
e.g. with the mu4e-maildirs-extension package) or another view mode
together with the minibuffer menu.