Fix handling of --text option in `msg-graphs' example
The previous implementation of `msg-graphs` called `mu:plots/mu:plot` with
output set to `#t` when the option `--text` was passed.  This caused an error in
the `string-append` call in `mu:plot` since the expression `(or output "dumb")`
evaluated to the symbol `#t` rather than a string.

This commit fixes this by making sure that `mu:plot` is called with `output` set
to `"dumb"` when the option `--text` is passed to `msg-graphs`, and with
`output` set to `"wxt"` (the default multi-platform graphical output device in
`gnuplot`) otherwise.  This is done by essentially making the code of
`msg-graphs` agree with the plotting code in the `guile/scripts` directory.