mu4e: Allow "washing" of URLs in message view
Add a new variable `mu4e-view-url-washers` which may contain a list of
functions to be invoked for replacing displayed text and (optionally)
link targets.

The motivation for this feature is that some corporate "advanced
thread protection" systems actually perform a man-in-the-middle-style
rewriting of links found in emails, in the process garbling the links
and making them unreadable, at least for plain-text emails; for HTML
mails, presumably only the link target is rewritten.

As an example, consider the following code, which will revert the
atrocities performed by Microsoft Outlook 365 on the links in your
mail:

  (defun sanitize-outlook-url (url)
    (let ((urlobj (url-generic-parse-url url)))
      (if (string= (url-host urlobj) "eur01.safelinks.protection.outlook.com")
          (let ((path+query (url-path-and-query urlobj)))
            (if (string= "/" (car path+query))
                (let ((extracted (cadr (assoc "url" (url-parse-query-string (cdr path+query))))))
                  (if extracted (list extracted extracted))))))))

  (add-to-list 'mu4e-view-url-washers #'sanitize-outlook-url)

The above code would replace *both* link target and displayed text,
thus removing the "protection". To keep being "protected", the
function could return just `extracted`, which will keep the actual
link target as-is, thus still going through the "safelinks"
redirection service.

Since this feature can lead to link targets differing from the
displayed URL, the `help-echo` property, which is shown on mouse-over,
will include the actual target of the link in this case.