manual: Deal with annoying misfeature of texinfo
Texinfo is of the opinion that every hyperlink has to be prefixed
by either "see" or "in".  If neither of these words appears there,
then it inserts "see" and it cannot be told not to do that.

This turns a correct sentence like:
  > Mu4e lets you define custom actions for messages in the <Headers view>.
into:
  > Mu4e lets you define custom actions for messages in the see <Headers view>.
The best compromise is
  > Mu4e lets you define custom actions for messages in <Headers view>.
which isn't correct but at least less messed up.

Alternatively one could rephrase every sentence that contains a link
to circumvent Texinfo's speech impairment.  Or one could replace each
link with a footnote and place the actual link in the footnote.